+++ demos/2D/demo2D04/demo2D04.sh
@@ -41,13 +41,19 @@
 # For additional information visit http://murphylab.web.cmu.edu or
 # send email to murphy@cmu.edu
 
-wget -nc --quiet http://www.cellorganizer.org/Downloads/v2.7/docker/demo2D04.tgz
+directory=../../../images/HeLa/2D/LAM
 
-mkdir -p ../../../images/HeLa/2D
+if [ ! -d "$directory" ]; then
 
-tar -xvf demo2D04.tgz -C ../../../images/HeLa/2D/
+	wget -nc --quiet http://www.cellorganizer.org/Downloads/v2.7/docker/demo2D04.tgz
 
-rm -f demo2D04.tgz
+	mkdir -p ../../../images/HeLa/2D/LAM
+
+	tar -xvf demo2D04.tgz -C ../../../images/HeLa/2D/LAM/
+
+	rm -f demo2D04.tgz
+
+fi
 
 echo -e "options.verbose = true;
 options.debug = false;
